Setup in-memory HSQL DB for Local Profile (default); Update DAO calls to use new base jdbc dao class with NamedParameterJdbcTemplate for simpler queries (more simplifying to come); use Long's intead of int/Integer for consistency and future-proofing. Repackaged mysql folder on src/main/resources

+++ src/main/java/svc/controllers/CourtController.java
@@ -30,7 +30,7 @@ public class CourtController {
 	
 	@ResponseBody
 	@RequestMapping(method = RequestMethod.GET, value="/{id}")
-	Court GetCourt(@PathVariable("id") Integer id) {
+	Court GetCourt(@PathVariable("id") Long id) {
 		if (id == null) {
 			LogSystem.LogEvent("Null id passed to controller");
 		}
